requirements.txt
attrs==24.3.0
blinker==1.7.0
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
dnspython==2.7.0
Flask==3.1.0
gunicorn==23.0.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
kaitaistruct==0.10
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==24.2
pyasn1==0.6.1
pycparser==2.22
pydivert==2.1.0
pymongo==4.10.1
pyOpenSSL==24.3.0
pyparsing==3.2.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.27.1
selenium-wire==5.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
zstandard==0.23.0
